+++ changelog.md
@@ -3,6 +3,7 @@
 ## v1.0.5 - [Not Released Yet]
 
 * Fixed https://github.com/APTrust/dart-runner/issues/18, which caused Dart Runner to create invalid paths in a bag's data directory under certain very specific conditions.
+* Fixed https://github.com/APTrust/dart-runner/issues/21, which caused Dart Runner to create invalid paths in a bag's manifest files when creating loose (untarred) bags.
 
 ## v1.0.4 - 2026-04-23